+++ cs/src/NativeMethods/LinkSimulator.cs
@@ -1,0 +1,13 @@
+// This file was automatically generated from header file
+using System;
+using System.Runtime.InteropServices;
+            
+namespace AUTD3Sharp.NativeMethods
+{
+    internal static class LinkSimulator
+    {
+        const string DLL = "autd3capi-link-simulator";
+
+        [DllImport(DLL, CharSet = CharSet.Ansi, BestFitMapping = false, ThrowOnUnmappableChar = true, CallingConvention = CallingConvention.Cdecl)] public static extern void AUTDLinkSimulator(out IntPtr @out, ushort port, string ipAddr);
+    }
+}