+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 2.4
+# 2.4.1
+
+- Add `extra::simulator::Simulator`
+- Rename `link::Emulator` to `link::Simulator`
+
+# 2.4.0
 
 - Add `GeometryViewer`
 - Improve performance of `link::SOEM` on Windows